+++ src/pool.h
@@ -18,7 +18,7 @@ typedef void (*POOL_worker)(cval obj, cval ctx);
 pool*	POOL_new();
 void	POOL_alloc(pool* p, int nr, POOL_worker f);
 bool	POOL_submit_job(pool* p, cval data, cval ctx);
-size_t	POOL_get_nr_workers(pool* p);
+int	POOL_get_nr_workers(pool* p);
 void	POOL_set_nr_workers(pool* p, int nr);
 void	POOL_free(pool* p);
